MeaningsText updation on delete and update enhanced

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/tasks/MeaningsTask.java
@@ -1,10 +1,10 @@
 package org.apache.atlas.repository.store.graph.v2.tasks;
 
 import org.apache.atlas.RequestContext;
-import org.apache.atlas.discovery.EntityDiscoveryService;
 import org.apache.atlas.exception.AtlasBaseException;
 import org.apache.atlas.exception.EntityNotFoundException;
 import org.apache.atlas.model.tasks.AtlasTask;
+import org.apache.atlas.repository.store.graph.v2.AtlasEntityStoreV2;
 import org.apache.atlas.repository.store.graph.v2.EntityGraphMapper;
 import org.apache.atlas.repository.store.graph.v2.preprocessor.glossary.TermPreProcessor;
 import org.apache.atlas.tasks.AbstractTask;
@@ -18,22 +18,25 @@ import java.util.Map;
 
 import static org.apache.atlas.model.tasks.AtlasTask.Status.COMPLETE;
 import static org.apache.atlas.model.tasks.AtlasTask.Status.FAILED;
+import static org.apache.atlas.repository.store.graph.v2.tasks.MeaningsTaskFactory.MEANINGS_TEXT_HARD_DELETE;
 
 public abstract class MeaningsTask extends AbstractTask {
     private static final Logger LOG = LoggerFactory.getLogger(MeaningsTask.class);
     protected static final String PARAM_ENTITY_GUID = "entityGuid";
-    protected static final String PARAM_ENTITY_QUALIFIED_NAME = "entityQname";
+    protected static final String PARAM_ENTITY_QUALIFIED_NAME = "entityQName";
     protected static final String PARAM_TERM_NAME = "termName";
-    protected final EntityDiscoveryService entityDiscovery;
+
     protected final EntityGraphMapper entityGraphMapper;
     protected final TermPreProcessor preprocessor;
+    protected final AtlasEntityStoreV2 entityStoreV2;
 
 
-    public MeaningsTask(AtlasTask task, EntityDiscoveryService entityDiscovery, EntityGraphMapper entityGraphMapper,TermPreProcessor preprocessor) {
+    public MeaningsTask(AtlasTask task, EntityGraphMapper entityGraphMapper,
+                        TermPreProcessor preprocessor, AtlasEntityStoreV2 entityStoreV2) {
         super(task);
-        this.entityDiscovery = entityDiscovery;
         this.entityGraphMapper = entityGraphMapper;
         this.preprocessor = preprocessor;
+        this.entityStoreV2 = entityStoreV2;
     }
 
     @Override
@@ -76,11 +79,21 @@ public abstract class MeaningsTask extends AbstractTask {
             put(PARAM_TERM_NAME, updateTerm);
         }};
     }
+    public static  Map<String,Object> toParameters(String termQName, String termGuid){
+        return new HashMap<String,Object>(){{
+            put(PARAM_ENTITY_QUALIFIED_NAME, termQName);
+            put(PARAM_ENTITY_GUID, termGuid);
+        }};
+    }
 
     protected void setStatus(AtlasTask.Status status) {
         super.setStatus(status);
         try {
-            entityGraphMapper.removePendingTaskFromEntity((String) getTaskDef().getParameters().get(PARAM_ENTITY_GUID), getTaskGuid());
+            if(MEANINGS_TEXT_HARD_DELETE.equals(getTaskType())){
+                LOG.info("Entity Vertex Deleted, No Need to remove pending task for: {} ",getTaskGuid());
+            }else {
+                entityGraphMapper.removePendingTaskFromEntity((String) getTaskDef().getParameters().get(PARAM_ENTITY_GUID), getTaskGuid());
+            }
         } catch (EntityNotFoundException  e) {
             LOG.error("Error updating associated element for: {}", getTaskGuid(), e);
         }
